+++ Assets/Scripts/WorldGen/WorldGen.cs
@@ -341,16 +341,16 @@ public static class WorldGen {
 		var WaterLayerElevation = new NativeArray<float>(staticState.Count, Allocator.TempJob);
 
 		Queue<int> openList = new Queue<int>();
-		Utils.MemsetArray<sbyte>(staticState.Count, default(JobHandle), state.Plate, -1).Complete();
+		Utils.MemsetArray<short>(staticState.Count, default(JobHandle), state.Plate, -1).Complete();
 		for (int i=0;i<worldGenData.NumPlates;i++) {
 			int plateStart = _random.Next(staticState.Count);
 			openList.Enqueue(plateStart);
-			state.Plate[plateStart] = (sbyte)i;
+			state.Plate[plateStart] = (short)i;
 		}
 		while (openList.Count > 0)
 		{
 			int index = openList.Dequeue();
-			sbyte p = state.Plate[index];
+			short p = state.Plate[index];
 			int neighborCount = StaticState.GetMaxNeighbors(index, staticState.Neighbors);
 			for (int i = 0; i < neighborCount; i++)
 			{
